pyproject.toml
[project]
name = "project-hms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mysql-connector-python>=9.5.0",
]
